Update README.md
Saddam crypter is a program that can obfuscate, manipulate, encrypt, and decrypt different types of malware codes. And it also has the ability to detect any kind of virus in a few minutes.

In this program, You will not find any kind of infections and you can easily download this program it is also a completely problem-free program.

By using this impressive program malware detection can be tested through this software with various spyware and malware detection program, including fyxm.net custom malware and spyware found in it.

Through this 100% fud crypter, You can easily compressed archives and combine multiple files into a single file to make them easier to transport or save on disk space.